fix(db): use `uint` for account ID

+++ Source/Starlight.SDK/Database/Models/Account.cs
@@ -2,7 +2,7 @@ namespace Starlight.SDK.Database.Models;
 
 public sealed class Account
 {
-    public long Id { get; set; }
+    public uint Id { get; set; }
     public string Username { get; set; } = string.Empty;
     public string Email { get; set; } = string.Empty;
     public string Password { get; set; } = string.Empty;
